+++ Engine/Player.cpp
@@ -2,25 +2,25 @@
 
 void Player::IsOutsideBoundries()
 {
-	const int right = x + width;
-	const int bottom = y + height;
+	const float right = x + float(width);
+	const float bottom = y + float(height);
 
 	if ( x < 0 )
 	{
 		x = 0;
 	}
-	else if ( right >= Graphics::ScreenWidth )
+	else if ( right >= float(Graphics::ScreenWidth) )
 	{
-		x = ( Graphics::ScreenWidth - 1 ) - width;
+		x = float( Graphics::ScreenWidth - 1 ) - float(width);
 	}
 
 	if ( y < 0 )
 	{
 		y = 0;
 	}
-	else if ( bottom >= Graphics::ScreenHeight )
+	else if ( bottom >= float(Graphics::ScreenHeight) )
 	{
-		y = ( Graphics::ScreenHeight - 1 ) - height;
+		y = float( Graphics::ScreenHeight - 1 ) - float(height);
 	}
 }
 
@@ -37,12 +37,12 @@ Player::Player()
 	y = startingY;
 }
 
-const int Player::GetX() const
+const float Player::GetX() const
 {
 	return x;
 }
 
-const int Player::GetY() const
+const float Player::GetY() const
 {
 	return y;
 }
@@ -84,5 +84,5 @@ void Player::UpdateInput( const Keyboard& kbd, float dt )
 
 void Player::DrawPlayer( Graphics & gfx ) const
 {
-	gfx.DrawRectDim( x, y, width, height, playerColor );
+	gfx.DrawRectDim( int(x), int(y), width, height, playerColor );
 }
